+++ src/app/core/components/main/main.component.html
@@ -5,7 +5,7 @@
     <mat-sidenav-content class="app-frame">
         <app-header (onToggleTheme)="toggleTheme()" (onToggleSidenav)="setSidenavState(true)"
             [isDark]="isDark"></app-header>
-        <router-outlet></router-outlet>
+            <router-outlet></router-outlet>
         <app-footer></app-footer>
     </mat-sidenav-content>
 </mat-sidenav-container>
