+++ src/app/modules/patterns/standalone/standalone.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit, inject } from '@angular/core';
+import { CoreModule } from '@app/core/core.module';
 import { NotificationService } from '@app/core/services/notification.service';
 import { CanResponsiveDirective } from '@app/shared/directives/can-responsive.directive';
 
@@ -7,6 +8,9 @@ import { CanResponsiveDirective } from '@app/shared/directives/can-responsive.di
   // signals: true, // this is available in Angular 17+
   selector: 'app-standalone',
   template: `<p>standalone works!</p>`,
+  imports: [
+    CoreModule,
+  ],
   hostDirectives: [
     {
       directive: CanResponsiveDirective,
